Enhetstest på Periode, Periode flyttet til egen fil.
